requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
decorator==5.2.1
discord.py==2.6.3
edge-tts==7.2.3
emoji==2.14.1
Flask==3.1.2
frozenlist==1.7.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.0.2
gspread_asyncio==2.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
msgpack==1.1.1
multidict==6.6.4
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
platformdirs==4.4.0
pooch==1.8.2
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyNaCl==1.5.0
regex==2025.9.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.15.3
soundfile==0.13.1
soxr==0.5.0.post1
StrEnum==0.4.15
tabulate==0.9.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
